MAINT: move KnownReparameterisation and ReparameterisationDict
